Fix filepath for NMR fileread
I found that the filepath for the NMR Splitpole files wasn't being reproduced correctly. It was used to load the Rmn directory to read the files in but then (at least for me), it looked in the current directory to try to find the Rmn files.

I have (in a fairly ugly manner) fixed this problem by introducing a new TString (FileToRead) which is used to construct the whole path to the Rmn files.