libMDM now dynamically decides whether or not to compile the FORTRAN RAYTRACE code.
If a fortran comiler is available AND if the cmake generator
is make ("Unix Makefile"), the target RAYTKIN1.F is added to the
list and included in libMDM.so.  Also, a preprocessor variable,
USE_RAYTRACE is defined. This is used in MDMTrace.cpp to decide
to import the functions from the fortran file.

If a fortran compiler is NOT available (or if the generator is
not make), RAYTKIN1.F is not added to the list and USE_RAYTRACE
is not defined. In MDMTrace.cpp, the #ifndef of USE_RAYTRACE prompts
the compilation of "fake" extern "C" routines which satisfy the
symbols needed to make a complete library (suymbols that normally
woud be found in the RAYTKIN1.F fortran source file).