Added code to properly handle sectors when ring+sector matching is turned off.
What the code does is take the sector to be a "fake" sector at the center of
the wedge.  So this way all relevant things (phi angle, impact angle, etc)
are calculated off the center of the wedge later on in the code, which is the
best we can do w/o using the actual sectors.